Send app_exception events at session finalize time
Consolidates the timing of sending app_exception events, and
includes NDK events as well, so CFU will be tracked properly.

Also exposes the asynchronous callback from the Analytics Receiver
as a global task chain which the uncaught exception handler can
wait on, so that we can make sure all events are sent successfully.